Move EncoderType and ScheduleType to settings.py
